# backend/api.yaml
...
openapi: 3.0.3
info:
  title: Katalyst Placement Portal API
  description: |
    Decoupled Flask JSON API for Katalyst V2.

    - Auth: `Authorization: Bearer <access_token>` (JWT)
    - Roles: `admin`, `student`, `company` (blacklisted users get 403)
    - UI wording uses "drives"; API resources use `job_positions`
    - Errors: `{ "error", "message", "status" }`
  version: 2.0.0
  contact:
    name: Deva Vasista
    email: 21f2000226@ds.study.iitm.ac.in
servers:
  - url: http://127.0.0.1:3000
    description: Local Flask API (`cd backend && python app.py`)
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token from POST /api/auth/login
    RefreshAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Refresh token from login (use only on /api/auth/refresh)
  schemas:
    Error:
      type: object
      required: [error, message, status]
      properties:
        error: { type: string, example: Bad Request }
        message: { type: string }
        status: { type: integer, example: 400 }
    User:
      type: object
      properties:
        id: { type: integer }
        username: { type: string }
        email: { type: string, format: email }
        role:
          type: string
          enum: [admin, student, company, blacklisted]
        created_at: { type: string, format: date-time }
        student: { $ref: "#/components/schemas/Student" }
        company: { $ref: "#/components/schemas/Company" }
    Student:
      type: object
      properties:
        id: { type: integer }
        user_id: { type: integer }
        full_name: { type: string }
        phone: { type: string, nullable: true }
        resume_path: { type: string, nullable: true }
        cgpa: { type: number, format: float, nullable: true }
        graduation_year: { type: integer, nullable: true }
        skills: { type: string, nullable: true }
        notify_deadlines: { type: boolean }
        created_at: { type: string, format: date-time }
        user: { $ref: "#/components/schemas/User" }
    Company:
      type: object
      properties:
        id: { type: integer }
        user_id: { type: integer }
        name: { type: string }
        website: { type: string, nullable: true }
        hr_contact: { type: string, nullable: true }
        industry: { type: string, nullable: true }
        approval_status:
          type: string
          enum: [pending, approved, rejected]
        created_at: { type: string, format: date-time }
        user: { $ref: "#/components/schemas/User" }
    JobPosition:
      type: object
      description: Placement drive (UI name)
      properties:
        id: { type: integer }
        company_id: { type: integer }
        title: { type: string }
        description: { type: string, nullable: true }
        requirements: { type: string, nullable: true }
        min_cgpa:
          type: number
          format: float
          nullable: true
          description: NULL means no CGPA cutoff
        salary_min: { type: integer, nullable: true }
        salary_max: { type: integer, nullable: true }
        deadline: { type: string, format: date-time, nullable: true }
        status:
          type: string
          enum: [active, closed]
        approval_status:
          type: string
          enum: [pending, approved, rejected]
        created_at: { type: string, format: date-time }
        company: { $ref: "#/components/schemas/Company" }
    Application:
      type: object
      properties:
        id: { type: integer }
        student_id: { type: integer }
        job_position_id: { type: integer }
        status:
          type: string
          enum: [applied, shortlisted, selected, rejected]
        interview_at: { type: string, format: date-time, nullable: true }
        interview_notes: { type: string, nullable: true }
        applied_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        student: { $ref: "#/components/schemas/Student" }
        job_position: { $ref: "#/components/schemas/JobPosition" }
    Placement:
      type: object
      properties:
        id: { type: integer }
        application_id: { type: integer }
        student_id: { type: integer }
        company_id: { type: integer }
        salary: { type: number, format: float, description: Package in LPA }
        joining_date: { type: string, format: date, nullable: true }
        placed_at: { type: string, format: date-time }
        created_at: { type: string, format: date-time }
        student_name: { type: string }
        company_name: { type: string }
        job_title: { type: string }
    JobPositionInput:
      type: object
      properties:
        title: { type: string }
        description: { type: string }
        requirements: { type: string }
        min_cgpa: { type: number, format: float, nullable: true }
        salary_min: { type: integer, nullable: true }
        salary_max: { type: integer, nullable: true }
        deadline:
          type: string
          description: YYYY-MM-DD or ISO datetime
      required: [title]
...
  responses:
    BadRequest:
      description: Invalid input
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: Missing or invalid JWT
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: Wrong role or blacklisted
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
paths:
...
  /api/admin/dashboard:
    get:
      tags: [Admin]
      summary: Admin dashboard aggregates
      security: [{ BearerAuth: [] }]
      parameters:
        - name: q
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Stats and entity lists
          content:
            application/json:
              schema:
                type: object
                properties:
                  stats: { type: object, additionalProperties: true }
                  pending_companies:
                    type: array
                    items: { $ref: "#/components/schemas/Company" }
                  pending_job_positions:
                    type: array
                    items: { $ref: "#/components/schemas/JobPosition" }
                  students:
                    type: array
                    items: { $ref: "#/components/schemas/Student" }
                  companies:
                    type: array
                    items: { $ref: "#/components/schemas/Company" }
                  job_positions:
                    type: array
                    items: { $ref: "#/components/schemas/JobPosition" }
                  recent_applications:
                    type: array
                    items: { $ref: "#/components/schemas/Application" }
                  search_q: { type: string }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /api/admin/placements:
    get:
      tags: [Admin]
      summary: List all confirmed placements
      security: [{ BearerAuth: [] }]
      responses:
        "200":
          description: Placements
          content:
            application/json:
              schema:
                type: object
                properties:
                  placements:
                    type: array
                    items: { $ref: "#/components/schemas/Placement" }
  /api/admin/students:
    post:
      tags: [Admin]
      summary: Create student account
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, email, password, full_name]
              properties:
                username: { type: string }
                email: { type: string }
                password: { type: string }
                full_name: { type: string }
                phone: { type: string }
                cgpa: { type: number }
                graduation_year: { type: integer }
                skills: { type: string }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
                  student: { $ref: "#/components/schemas/Student" }
        "400": { $ref: "#/components/responses/BadRequest" }
  /api/admin/students/{student_id}:
    get:
      tags: [Admin]
      summary: Student detail and applications
      security: [{ BearerAuth: [] }]
      parameters:
        - name: student_id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Detail
          content:
            application/json:
              schema:
                type: object
                properties:
                  student: { $ref: "#/components/schemas/Student" }
                  applications:
                    type: array
                    items: { $ref: "#/components/schemas/Application" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [Admin]
      summary: Delete student and linked user
      security: [{ BearerAuth: [] }]
      parameters:
        - name: student_id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Deleted
        "404": { $ref: "#/components/responses/NotFound" }
  /api/admin/students/{student_id}/blacklist:
    post:
      tags: [Admin]
      summary: Blacklist student
      security: [{ BearerAuth: [] }]
      parameters:
        - name: student_id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Blacklisted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
                  student: { $ref: "#/components/schemas/Student" }
  /api/admin/students/{student_id}/unblacklist:
    post:
      tags: [Admin]
      summary: Re-enable blacklisted student
      security: [{ BearerAuth: [] }]
      parameters:
        - name: student_id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Re-enabled
        "400": { $ref: "#/components/responses/BadRequest" }
  /api/admin/companies:
    post:
      tags: [Admin]
      summary: Create company (auto-approved)
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, email, password, company_name]
              properties:
                username: { type: string }
                email: { type: string }
                password: { type: string }
                company_name: { type: string }
                website: { type: string }
                hr_contact: { type: string }
                industry: { type: string }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
                  company: { $ref: "#/components/schemas/Company" }
  /api/admin/companies/{company_id}:
    get:
      tags: [Admin]
      summary: Company detail, drives, and light stats
      security: [{ BearerAuth: [] }]
      parameters:
        - name: company_id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Detail
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [Admin]
      summary: Delete company and linked user
      security: [{ BearerAuth: [] }]
      parameters:
        - name: company_id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Deleted
  /api/admin/companies/{company_id}/approve:
    post:
      tags: [Admin]
      summary: Approve company
      security: [{ BearerAuth: [] }]
      parameters:
        - name: company_id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
                  company: { $ref: "#/components/schemas/Company" }
  /api/admin/companies/{company_id}/reject:
    post:
      tags: [Admin]
      summary: Reject company
      security: [{ BearerAuth: [] }]
      parameters:
        - name: company_id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Rejected
  /api/admin/companies/{company_id}/blacklist:
    post:
      tags: [Admin]
      summary: Blacklist company user
      security: [{ BearerAuth: [] }]
      parameters:
        - name: company_id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Blacklisted
  /api/admin/companies/{company_id}/unblacklist:
    post:
      tags: [Admin]
      summary: Re-enable company (sets approval pending)
      security: [{ BearerAuth: [] }]
      parameters:
        - name: company_id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Re-enabled
        "400": { $ref: "#/components/responses/BadRequest" }
  /api/admin/job-positions:
    post:
      tags: [Admin]
      summary: Create drive for a company (auto-approved)
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/JobPositionInput"
                - type: object
                  required: [company_id, title]
                  properties:
                    company_id: { type: integer }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
                  job_position: { $ref: "#/components/schemas/JobPosition" }
  /api/admin/job-positions/{job_id}:
    get:
      tags: [Admin]
      summary: Drive detail and applicants
      security: [{ BearerAuth: [] }]
      parameters:
        - name: job_id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Detail with stats and applications
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [Admin]
      summary: Delete drive
      security: [{ BearerAuth: [] }]
      parameters:
        - name: job_id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Deleted
  /api/admin/job-positions/{job_id}/approve:
    post:
      tags: [Admin]
      summary: Approve drive
      security: [{ BearerAuth: [] }]
      parameters:
        - name: job_id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
                  job_position: { $ref: "#/components/schemas/JobPosition" }
  /api/admin/job-positions/{job_id}/reject:
    post:
      tags: [Admin]
      summary: Reject drive
      security: [{ BearerAuth: [] }]
      parameters:
        - name: job_id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Rejected

  # ---------- Company ----------
...